Workflow tests for Pickup and Delivery.

See merge request lego-front/auto-tests!166

+++ src/test/java/com/leroy/magmobile/api/data/sales/orders/OrderProductData.java
@@ -17,7 +17,7 @@ public class OrderProductData extends BaseProductOrderData {
     private String longTail;
     private String stock;
     private String vat;
-    private Integer confirmedQuantity;
+    private Double confirmedQuantity;
     private String pointOfSale;
     private String pointOfOrigin;
